pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "docs_agent"
version = "0.0.1"
description = "The docs agent finds, pulls and learns the documentation for your tools, libraries and programming languages so that you don't have to. Instead of switching tabs to Google (or your search engine of choice), ask docs about the tools in your project."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.4.1",
    "docopt-ng>=0.9.0",
    "ollama>=0.6.1",
    "platformdirs>=4.5.1",
    "python-dotenv>=1.2.1",
]

[project.scripts]
docs = "docs_agent.__main__:main"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.14.14",
]
